Cruds transfe a ih d

+++ src/Controller/CustomerCardController.php
@@ -5,9 +5,11 @@ namespace App\Controller;
 use App\Entity\Comment;
 use App\Entity\CustomerCard;
 use App\Entity\StatusHistory;
+use App\Entity\TransferArrival;
 use App\Entity\User;
 use App\Form\CommentType;
 use App\Form\CustomerCardType;
+use App\Form\CustomerCardNewType;
 use App\Repository\AgencyRepository;
 use App\Repository\AirportHotelRepository;
 use App\Repository\CommentRepository;
@@ -27,6 +29,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 class CustomerCardController extends AbstractController
 {
@@ -466,22 +469,58 @@ class CustomerCardController extends AbstractController
     }
 
 
-    #[Route('/new', name: 'app_customer_card_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, CustomerCardRepository $customerCardRepository): Response
+    #[Route('team-manager/customer/card/new', name: 'app_customer_card_new', methods: ['GET', 'POST'])]
+    public function new(Request $request, 
+                        CustomerCardRepository $customerCardRepository,
+                        AirportHotelRepository $airportHotelRepository,
+                        TransferArrivalRepository $transferArrivalRepository
+                        ): Response
     {
+
+
+
         $customerCard = new CustomerCard();
-        $form = $this->createForm(CustomerCardType::class, $customerCard);
+        $form = $this->createForm(CustomerCardNewType::class, $customerCard);
         $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $customerCardRepository->save($customerCard, true);
+        // récupérer les airports
+        $airports = $airportHotelRepository->findBy(['isAirport' => true]);
+        // récupérer les hotels
+        $hotels = $airportHotelRepository->findBy(['isAirport' => false]);
 
+        if ($form->isSubmitted() && $form->isValid()) {
+            $customerCard->setStatusUpdatedBy($this->getUser());
+
+            // mettre a jour le meeting grace a l'arrivée
+            $meetingDate = new DateTimeImmutable($request->request->get('date'));
+            $meetingDateFormat = new DateTimeImmutable($meetingDate->format('Y-m-d 00:01'));
+            $meetingDateHour = $meetingDateFormat->modify('+1 day');
+            $customerCard->setMeetingAt( $meetingDateHour);  
+
+            $airport = $airportHotelRepository->find($request->request->get('fromStart'));
+            $hotel = $airportHotelRepository->find($request->request->get('toArrival'));
+
+            $arrival = new TransferArrival();
+            $arrival->setServiceNumber($request->request->get('serviceNumber'));
+            $arrival->setFlightNumber($request->request->get('flightNumber'));
+            $arrival->setIsCollective($request->request->get('isCollective'));
+            $arrival->setDate(new DateTimeImmutable($request->request->get('date')));
+            $arrival->setHour(new DateTimeImmutable($request->request->get('hour')));
+            $arrival->setDateHour(new DateTimeImmutable($request->request->get('dateHour')));
+            $arrival->setFromStart($airport);
+            $arrival->setToArrival($hotel);
+            $arrival->setCustomerCard($customerCard);
+
+            $customerCardRepository->save($customerCard, true);
+            $transferArrivalRepository->save($arrival, true);
             return $this->redirectToRoute('app_customer_card_index', [], Response::HTTP_SEE_OTHER);
         }
 
         return $this->render('customer_card/new.html.twig', [
             'customer_card' => $customerCard,
             'form' => $form,
+            'airports' => $airports,
+            'hotels' => $hotels,
         ]);
     }
 
@@ -489,7 +528,6 @@ class CustomerCardController extends AbstractController
     #[Route('customer/card/{id}', name: 'app_customer_card_show', methods: ['GET' , 'POST'])]
     public function show(CustomerCard $customerCard, Request $request, CommentRepository $commentRepository, UserRepository $userRepository): Response
     {
-
 
 
         $user = $this->getUser();
@@ -652,7 +690,8 @@ class CustomerCardController extends AbstractController
 
             $customerCardRepository->save($customerCard, true);
 
-            return $this->redirectToRoute('app_customer_card_index', [
+            return $this->redirectToRoute('app_customer_card_show', [
+                'id' => $customerCard->getId(),
                 'customerPresence'=> $customerPresence,
                 'dateStart'=> $dateStart,
                 'dateEnd'=> $dateEnd,
@@ -667,15 +706,46 @@ class CustomerCardController extends AbstractController
                 , Response::HTTP_SEE_OTHER);
         }
 
-        return $this->renderForm('customer_card/edit.html.twig', [
+        return $this->render('customer_card/edit.html.twig', [
             'customer_card' => $customerCard,
             'form' => $form,
         ]);
     }
 
     #[Route('/{id}', name: 'app_customer_card_delete', methods: ['POST'])]
-    public function delete(Request $request, CustomerCard $customerCard, CustomerCardRepository $customerCardRepository): Response
+    public function delete(Request $request, 
+                            CustomerCard $customerCard, CustomerCardRepository $customerCardRepository, 
+                            TransferArrivalRepository $transferArrivalRepository,
+                            TranslatorInterface $translator): Response
     {
+
+
+        
+        if ($customerCard->getTransferInterHotels()->count() > 0) {
+            // vous ne pouvez pas supprimer car ce client a des interHotels. Supprimer les départs avant 
+            $this->addFlash(
+                'warning',
+                'You can\'t delete this client card because there is an interHotel associated. Please remove the inter hotel first'
+            );
+            return $this->redirectToRoute('app_customer_card_show', ['id' => $customerCard->getId()], Response::HTTP_SEE_OTHER);
+        }
+        if ($customerCard->getTransferDeparture()->count() > 0) {
+            // vous ne pouvez pas supprimer car ce client a des departs. Supprimer les départs avant 
+            $this->addFlash(
+                'warning',
+                'You can\'t delete this client card because there is a deperture associated. Please remove the departure first'
+            );
+            return $this->redirectToRoute('app_customer_card_show', ['id' => $customerCard->getId()], Response::HTTP_SEE_OTHER);
+        }
+
+        if ($customerCard->getTransferArrivals()->count() > 0) {
+            // récupérer les arrivées et les supprimer 1 a 1
+            foreach ($customerCard->getTransferArrivals() as $transfer) {
+               $currentTransferArrival = $transferArrivalRepository->find($transfer);
+               $transferArrivalRepository->remove($currentTransferArrival, true);
+            } 
+        }
+
         if ($this->isCsrfTokenValid('delete'.$customerCard->getId(), $request->request->get('_token'))) {
             $customerCardRepository->remove($customerCard, true);
         }
